refactor: split deployment source handling
