+++ fastmcp_slim/fastmcp/cli/deploy/source_bundle.py
@@ -1,49 +1,27 @@
-"""Resolve and archive local source for a Horizon deployment."""
+"""Create deterministic archives for Horizon deployment source."""
 
 from __future__ import annotations
 
-import ast
 import gzip
 import hashlib
 import io
-import json
 import os
 import stat
 import tarfile
-import tokenize
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from pathlib import Path
 
 import pathspec
-from pydantic import ValidationError
 
-from fastmcp.utilities.mcp_server_config import MCPServerConfig
-from fastmcp.utilities.mcp_server_config.v1.sources.filesystem import FileSystemSource
+from fastmcp.cli.deploy.source import GENERATED_REQUIREMENTS_PATH, DeploySource
+from fastmcp.cli.deploy.source_paths import (
+    SourceInvalidError,
+    is_fixed_exclusion,
+    require_included_path,
+    validate_symlink,
+)
 
 MAX_SOURCE_UPLOAD_BYTES = 250 * 1024 * 1024
-GENERATED_REQUIREMENTS_PATH = ".fastmcp-deploy-requirements.txt"
-
-_EXCLUDED_DIRECTORY_NAMES = {
-    ".fastmcp",
-    ".git",
-    ".mypy_cache",
-    ".nox",
-    ".pytest_cache",
-    ".ruff_cache",
-    ".tox",
-    ".venv",
-    "__pycache__",
-    "env",
-    "venv",
-    "virtualenv",
-}
-_EXCLUDED_FILE_NAMES = {".DS_Store"}
-_EXCLUDED_FILE_SUFFIXES = {".pyc", ".pyo"}
-_COMMON_ENTRYPOINTS = ("mcp", "server", "app")
-
-
-class SourceInvalidError(ValueError):
-    """The local deployment source is invalid or unsafe."""
 
 
 class ArchiveTooLargeError(ValueError):
@@ -66,17 +44,6 @@ class _CompressedSizeWriter:
 
 
 @dataclass(frozen=True)
-class DeploySource:
-    """A resolved local source before archive creation."""
-
-    source_root: Path
-    entrypoint: str
-    dependency_path: str | None
-    _generated_requirements: str | None = field(default=None, repr=False)
-    _required_paths: tuple[Path, ...] = field(default=(), repr=False)
-
-
-@dataclass(frozen=True)
 class SourceBundle:
     """An immutable archive ready for source upload."""
 
@@ -85,106 +52,6 @@ class SourceBundle:
     checksum_sha256: str
     entrypoint: str
     dependency_path: str | None
-
-
-@dataclass(frozen=True)
-class _DependencyInput:
-    path: str | None
-    generated_requirements: str | None
-    required_paths: tuple[Path, ...]
-
-
-class _ModuleBindingVisitor(ast.NodeVisitor):
-    def __init__(self) -> None:
-        self.names: set[str] = set()
-
-    def visit_Name(self, node: ast.Name) -> None:
-        if isinstance(node.ctx, ast.Store):
-            self.names.add(node.id)
-
-    def visit_FunctionDef(self, node: ast.FunctionDef) -> None:
-        self.names.add(node.name)
-
-    def visit_AsyncFunctionDef(self, node: ast.AsyncFunctionDef) -> None:
-        self.names.add(node.name)
-
-    def visit_ClassDef(self, node: ast.ClassDef) -> None:
-        pass
-
-    def visit_Import(self, node: ast.Import) -> None:
-        for alias in node.names:
-            self.names.add(alias.asname or alias.name.partition(".")[0])
-
-    def visit_ImportFrom(self, node: ast.ImportFrom) -> None:
-        for alias in node.names:
-            if alias.name != "*":
-                self.names.add(alias.asname or alias.name)
-
-    def visit_ListComp(self, node: ast.ListComp) -> None:
-        pass
-
-    def visit_SetComp(self, node: ast.SetComp) -> None:
-        pass
-
-    def visit_DictComp(self, node: ast.DictComp) -> None:
-        pass
-
-    def visit_GeneratorExp(self, node: ast.GeneratorExp) -> None:
-        pass
-
-    def visit_Lambda(self, node: ast.Lambda) -> None:
-        pass
-
-
-async def resolve_deploy_source(server_spec: str | None) -> DeploySource:
-    """Resolve a FastMCP server input without applying runtime settings."""
-    config, source_root = _load_config(server_spec)
-    source_path = _resolve_path(
-        source_root,
-        config.source.path,
-        label="Server source",
-        file=True,
-    )
-    _require_included_path(source_root, source_path, label="Server source")
-
-    if config.source.entrypoint and ":" in config.source.entrypoint:
-        raise SourceInvalidError(
-            "Deployment entrypoints must use a file and one object name"
-        )
-
-    object_name = _resolve_entrypoint(source_path, config.source.entrypoint)
-    relative_source = source_path.relative_to(source_root).as_posix()
-    dependency = _resolve_dependencies(config, source_root, source_path)
-    return DeploySource(
-        source_root=source_root,
-        entrypoint=f"{relative_source}:{object_name}",
-        dependency_path=dependency.path,
-        _generated_requirements=dependency.generated_requirements,
-        _required_paths=(source_path, *dependency.required_paths),
-    )
-
-
-def _resolve_entrypoint(source_path: Path, explicit_entrypoint: str | None) -> str:
-    if explicit_entrypoint:
-        return explicit_entrypoint
-
-    try:
-        with tokenize.open(source_path) as source_file:
-            source_text = source_file.read()
-        module = ast.parse(source_text, filename=str(source_path))
-    except (OSError, SyntaxError, UnicodeError) as exc:
-        raise SourceInvalidError(
-            f"The server source could not be parsed: {exc}"
-        ) from exc
-
-    bindings = _ModuleBindingVisitor()
-    bindings.visit(module)
-    for name in _COMMON_ENTRYPOINTS:
-        if name in bindings.names:
-            return name
-    raise SourceInvalidError(
-        "No server object named mcp, server, or app was found; provide an entrypoint"
-    )
 
 
 def create_source_bundle(
@@ -210,16 +77,16 @@ def create_source_bundle(
     entries = _archive_entries(
         source_root,
         archive_path,
-        required_paths=source._required_paths,
+        required_paths=source.required_paths,
     )
-    if source._generated_requirements is not None:
+    if source.generated_requirements is not None:
         generated_path = source_root / GENERATED_REQUIREMENTS_PATH
         if os.path.lexists(generated_path):
             raise SourceInvalidError(
                 f"The generated dependency path is reserved: {GENERATED_REQUIREMENTS_PATH}"
             )
         entries.append(
-            (GENERATED_REQUIREMENTS_PATH, source._generated_requirements.encode())
+            (GENERATED_REQUIREMENTS_PATH, source.generated_requirements.encode())
         )
     entries.sort(key=lambda entry: entry[0])
 
@@ -259,221 +126,6 @@ def create_source_bundle(
     )
 
 
-def _load_config(server_spec: str | None) -> tuple[MCPServerConfig, Path]:
-    if server_spec is None:
-        config_path = MCPServerConfig.find_config()
-        if config_path is None:
-            raise SourceInvalidError(
-                "Provide a server input or add fastmcp.json to the current directory"
-            )
-        return _read_config(config_path), config_path.parent.resolve()
-
-    if server_spec.endswith(".json"):
-        config_path = Path(server_spec).expanduser().absolute()
-        return _read_config(config_path), config_path.parent.resolve()
-
-    return (
-        MCPServerConfig(source=FileSystemSource(path=server_spec)),
-        Path.cwd().resolve(),
-    )
-
-
-def _read_config(config_path: Path) -> MCPServerConfig:
-    try:
-        return MCPServerConfig.from_file(config_path)
-    except (FileNotFoundError, json.JSONDecodeError, ValidationError) as exc:
-        raise SourceInvalidError(
-            f"The FastMCP configuration is invalid: {config_path}"
-        ) from exc
-
-
-def _resolve_dependencies(
-    config: MCPServerConfig,
-    source_root: Path,
-    source_path: Path,
-) -> _DependencyInput:
-    environment = config.environment
-    requirements = (
-        _resolve_path(
-            source_root,
-            environment.requirements,
-            label="Requirements file",
-            file=True,
-        )
-        if environment.requirements is not None
-        else None
-    )
-    project = (
-        _resolve_path(
-            source_root,
-            environment.project,
-            label="Environment project",
-            directory=True,
-        )
-        if environment.project is not None
-        else None
-    )
-    editable = tuple(
-        _resolve_path(
-            source_root,
-            path,
-            label="Editable dependency",
-            directory=True,
-        )
-        for path in (environment.editable or [])
-    )
-
-    for label, path in (
-        ("Requirements file", requirements),
-        ("Environment project", project),
-        *(("Editable dependency", path) for path in editable),
-    ):
-        if path is not None:
-            _require_included_path(source_root, path, label=label)
-
-    dependencies = sorted(set(environment.dependencies or []))
-    needs_generated = bool(
-        dependencies or editable or (project is not None and requirements is not None)
-    )
-    if needs_generated:
-        lines: list[str] = []
-        required_paths: list[Path] = []
-        if requirements is not None:
-            lines.append(f"-r {_relative_path(source_root, requirements)}")
-            required_paths.append(requirements)
-        if project is not None:
-            lines.append(f"-e {_relative_path(source_root, project)}")
-            required_paths.append(project)
-            project_config = project / "pyproject.toml"
-            if project_config.is_file():
-                required_paths.append(project_config)
-        for path in sorted(
-            editable, key=lambda item: _relative_path(source_root, item)
-        ):
-            lines.append(f"-e {_relative_path(source_root, path)}")
-            required_paths.append(path)
-            editable_config = path / "pyproject.toml"
-            if editable_config.is_file():
-                required_paths.append(editable_config)
-        lines.extend(dependencies)
-        return _DependencyInput(
-            path=GENERATED_REQUIREMENTS_PATH,
-            generated_requirements="\n".join(lines) + "\n",
-            required_paths=tuple(required_paths),
-        )
-
-    if requirements is not None:
-        return _DependencyInput(
-            path=_relative_path(source_root, requirements),
-            generated_requirements=None,
-            required_paths=(requirements,),
-        )
-
-    if project is not None:
-        dependency = _dependency_input_in(
-            source_root,
-            project,
-            include_requirements=False,
-        )
-        if dependency is not None:
-            return _DependencyInput(
-                path=dependency.path,
-                generated_requirements=None,
-                required_paths=(project, *dependency.required_paths),
-            )
-        raise SourceInvalidError(
-            "The environment project has no pyproject.toml dependency file"
-        )
-
-    return _detect_dependency_input(source_root, source_path.parent)
-
-
-def _detect_dependency_input(source_root: Path, start: Path) -> _DependencyInput:
-    current = start
-    while True:
-        dependency = _dependency_input_in(source_root, current)
-        if dependency is not None:
-            return dependency
-        if current == source_root:
-            return _DependencyInput(
-                path=None,
-                generated_requirements=None,
-                required_paths=(),
-            )
-        current = current.parent
-
-
-def _dependency_input_in(
-    source_root: Path,
-    directory: Path,
-    *,
-    include_requirements: bool = True,
-) -> _DependencyInput | None:
-    requirements = directory / "requirements.txt"
-    if include_requirements and requirements.is_file():
-        return _DependencyInput(
-            path=_relative_path(source_root, requirements),
-            generated_requirements=None,
-            required_paths=(requirements,),
-        )
-
-    uv_lock = directory / "uv.lock"
-    pyproject = directory / "pyproject.toml"
-    if uv_lock.is_file() and pyproject.is_file():
-        return _DependencyInput(
-            path=_relative_path(source_root, uv_lock),
-            generated_requirements=None,
-            required_paths=(uv_lock, pyproject),
-        )
-    if pyproject.is_file():
-        return _DependencyInput(
-            path=_relative_path(source_root, pyproject),
-            generated_requirements=None,
-            required_paths=(pyproject,),
-        )
-    return None
-
-
-def _resolve_path(
-    source_root: Path,
-    value: str | Path,
-    *,
-    label: str,
-    file: bool = False,
-    directory: bool = False,
-) -> Path:
-    requested = Path(value).expanduser()
-    candidate = Path(
-        os.path.abspath(
-            requested if requested.is_absolute() else source_root / requested
-        )
-    )
-    try:
-        resolved = candidate.resolve(strict=True)
-    except (OSError, RuntimeError) as exc:
-        raise SourceInvalidError(f"{label} does not exist: {value}") from exc
-    try:
-        resolved.relative_to(source_root)
-    except ValueError:
-        raise SourceInvalidError(f"{label} leaves the source root: {value}") from None
-    if file and not candidate.is_file():
-        raise SourceInvalidError(f"{label} is not a file: {value}")
-    if directory and not candidate.is_dir():
-        raise SourceInvalidError(f"{label} is not a directory: {value}")
-    return candidate
-
-
-def _require_included_path(source_root: Path, path: Path, *, label: str) -> None:
-    relative = path.relative_to(source_root)
-    if _is_fixed_exclusion(relative):
-        raise SourceInvalidError(f"{label} is excluded from deployment: {relative}")
-
-
-def _relative_path(source_root: Path, path: Path) -> str:
-    relative = path.relative_to(source_root).as_posix()
-    return relative or "."
-
-
 def _archive_entries(
     source_root: Path,
     archive_path: Path,
@@ -503,7 +155,7 @@ def _archive_entries(
         for name in sorted(directory_names):
             path = current / name
             relative = path.relative_to(source_root)
-            if _is_fixed_exclusion(relative):
+            if is_fixed_exclusion(relative):
                 continue
             if _is_gitignored(path, rules, directory=True):
                 if any(
@@ -514,7 +166,7 @@ def _archive_entries(
                     )
                 continue
             if path.is_symlink():
-                _validate_symlink(source_root, path)
+                validate_symlink(source_root, path)
                 entries[relative.as_posix()] = path
             else:
                 traversable_directories.append(name)
@@ -526,12 +178,12 @@ def _archive_entries(
             if path.absolute() == archive_path:
                 continue
             relative = path.relative_to(source_root)
-            if _is_fixed_exclusion(relative):
+            if is_fixed_exclusion(relative):
                 continue
             if path.absolute() not in required and _is_gitignored(path, rules):
                 continue
             if path.is_symlink():
-                _validate_symlink(source_root, path)
+                validate_symlink(source_root, path)
             elif not path.is_file():
                 continue
             entries[relative.as_posix()] = path
@@ -557,10 +209,10 @@ def _expand_required_paths(
         path = pending.pop().absolute()
         if path in required:
             continue
-        _require_included_path(source_root, path, label="Required source path")
+        require_included_path(source_root, path, label="Required source path")
         required.add(path)
         if path.is_symlink():
-            pending.append(_validate_symlink(source_root, path))
+            pending.append(validate_symlink(source_root, path))
     return required
 
 
@@ -572,7 +224,7 @@ def _read_gitignore(
     if not gitignore.is_file():
         return None
     if gitignore.is_symlink():
-        _validate_symlink(source_root, gitignore)
+        validate_symlink(source_root, gitignore)
     try:
         lines = gitignore.read_text(encoding="utf-8").splitlines()
     except (OSError, UnicodeError) as exc:
@@ -598,32 +250,6 @@ def _is_gitignored(
     return ignored
 
 
-def _is_fixed_exclusion(relative: Path) -> bool:
-    parts = relative.parts
-    if any(part in _EXCLUDED_DIRECTORY_NAMES for part in parts[:-1]):
-        return True
-    name = relative.name
-    return (
-        name in _EXCLUDED_DIRECTORY_NAMES
-        or name in _EXCLUDED_FILE_NAMES
-        or name == ".env"
-        or name.startswith(".env.")
-        or relative.suffix in _EXCLUDED_FILE_SUFFIXES
-    )
-
-
-def _validate_symlink(source_root: Path, path: Path) -> Path:
-    try:
-        target = path.resolve(strict=True)
-        target.relative_to(source_root)
-    except (OSError, RuntimeError, ValueError):
-        relative = path.relative_to(source_root)
-        raise SourceInvalidError(
-            f"Symbolic link leaves the source root: {relative}"
-        ) from None
-    return target
-
-
 def _add_archive_entry(
     archive: tarfile.TarFile,
     name: str,
@@ -645,7 +271,7 @@ def _add_archive_entry(
         return
 
     if entry_source.is_symlink():
-        target = _validate_symlink(source_root, entry_source)
+        target = validate_symlink(source_root, entry_source)
         info.type = tarfile.SYMTYPE
         info.mode = 0o777
         info.linkname = os.path.relpath(target, entry_source.parent).replace(
